8153992: Some hotspot tests fail on compact2 due to an unnecessary test library dependency
Removing unnecessary dependency on java.management from the ProcessTool.getProcessId()

+++ hotspot/test/runtime/NMT/VirtualAllocCommitUncommitRecommit.java
@@ -48,7 +48,7 @@ public class VirtualAllocCommitUncommitRecommit {
         long reserveSize = 4 * 1024 * 1024; // 4096KB
         long addr;
 
-        String pid = Integer.toString(ProcessTools.getProcessId());
+        String pid = Long.toString(ProcessTools.getProcessId());
         ProcessBuilder pb = new ProcessBuilder();
 
         // reserve
